requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docxcompose==1.4.0
docxtpl==0.16.7
Flask==2.3.2
frozenlist==1.4.1
gunicorn==20.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.1
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python==8.0.33
openai==0.27.8
protobuf==3.20.3
pyTelegramBotAPI==4.12.0
python-docx==1.1.2
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
six==1.16.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
